resrc: Moved unit tests out of resrc.c and into test/tresrc.c
Follows this proposal in flux-core:
https://github.com/flux-framework/flux-core/pull/145#issuecomment-76024889
