+++ resrc/test/Makefile
@@ -10,16 +10,20 @@ LUA_CPATH = $(abs_topdir)/rdl/?.so;$(FLUX_SRCDIR)/src/bindings/lua/.libs/?.so;;;
 
 export CFLAGS TESTRESRC_INPUT_FILE LUA_PATH LUA_CPATH
 
-LDADD_CLI = $(FLUX_SRCDIR)/src/common/libutil/.libs/libutil.a
+LDADD_CLI = $(FLUX_SRCDIR)/src/common/libutil/.libs/libutil.a \
+            $(FLUX_SRCDIR)/src/common/libtap/.libs/libtap.a
+
+test_cppflags = \
+	-I$(FLUX_SRCDIR)/src/common/libtap
 
 BUILD = tresrc
 
 all: $(BUILD)
 
 tresrc: tresrc.o
-	$(CC) -o $@ $^ $(LIBS) $(LDADD_CLI)
+	$(CC) $(test_cppflags) -o $@ $^ $(LIBS) $(LDADD_CLI)
 
-check:
+check: tresrc
 	./tresrc
 
 clean:
