+++ resrc/test/tresrc.c
@@ -22,26 +22,21 @@
  *  See also:  http://www.gnu.org/licenses/
 \*****************************************************************************/
 
-#include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <stdbool.h>
 #include <string.h>
-#include <uuid/uuid.h>
+#include <sys/time.h>
 #include <czmq.h>
 
 #include "../resrc.h"
 #include "../resrc_tree.h"
-#include "src/common/libutil/jsonutil.h"
-#include "src/common/libutil/xzmalloc.h"
+#include "src/common/libtap/tap.h"
 
-#include <sys/time.h>
-#include <sys/types.h>
 
 static struct timeval start_time;
 
 void init_time() {
-              gettimeofday(&start_time, NULL);
+    gettimeofday(&start_time, NULL);
 }
 
 u_int64_t get_time() {
@@ -51,93 +46,166 @@ u_int64_t get_time() {
         + (t.tv_usec - start_time.tv_usec);
 }
 
-int main (int argc, char** argv)
+
+int main (int argc, char *argv[])
 {
     char *resrc_id = NULL;
     const char *filename = argv[1];
     int found = 0;
-    JSON child_sock = NULL;
+    int rc = 0;
+    int verbose = 0;
+    JSON ja = NULL;
+    /* JSON jtago = NULL;  /\* json tag object *\/ */
     JSON child_core = NULL;
+    JSON child_sock = NULL;
     JSON o = NULL;
     JSON req_res = NULL;
-    resrc_t *resrc = NULL;
-    resrc_tree_t *resrc_tree = NULL;
-    resources_t *resrcs;
     resource_list_t *found_res = resrc_new_id_list ();
+    resources_t *resrcs = NULL;
+    resrc_t *resrc = NULL;
+    resrc_t *sample_resrc = NULL;
+    resrc_tree_t *resrc_tree = NULL;
+    zlist_t *resrc_tree_list = zlist_new ();
 
+    plan (14);
     if (filename == NULL || *filename == '\0')
         filename = getenv ("TESTRESRC_INPUT_FILE");
 
+    ok ((filename != NULL), "valid resource file name");
+    ok ((access (filename, F_OK) == 0), "resoure file exists");
+    ok ((access (filename, R_OK) == 0), "resoure file readable");
+
     init_time();
     resrcs = resrc_generate_resources (filename, "default");
-    printf("resource generation took: %lf\n", ((double)get_time())/1000000);
-    printf ("starting\n");
-    resrc_print_resources (resrcs);
-    printf ("end of resources\n");
 
-    printf ("printing resource tree\n");
+    ok ((resrcs != NULL), "resource generation took: %lf",
+        ((double)get_time())/1000000);
+    if (!resrcs)
+        goto ret;
+
+    if (verbose) {
+        printf ("Listing flat resources:\n");
+        resrc_print_resources (resrcs);
+        printf ("End of flat resources\n");
+    }
+
     resrc = zhash_lookup ((zhash_t *)resrcs, "cluster.0");
-    if (resrc)
-        resrc_tree_print (resrc_phys_tree (resrc));
-    else
-        printf ("Failed to find cluster.0 resource\n");
-    printf ("end of resource tree\n");
+    ok ((resrc != NULL), "cluster.0 tree head found");
+    if (!resrc)
+        goto ret;
 
+    resrc_tree = resrc_phys_tree (resrc);
+    ok ((resrc_tree != NULL), "resource tree valid");
+    if (!resrc_tree)
+        goto ret;
+
+    if (verbose) {
+        printf ("Listing resource tree\n");
+        resrc_tree_print (resrc_tree);
+        printf ("End of resource tree\n");
+    }
+    zlist_append (resrc_tree_list, resrc_tree);
+
+    /*
+     *  Build a resource composite to search for
+     */
     child_core = Jnew ();
     Jadd_str (child_core, "type", "core");
     Jadd_int (child_core, "req_qty", 6);
 
+    ja = Jnew_ar ();
+    Jadd_ar_obj(ja, child_core);
+
     child_sock = Jnew ();
     Jadd_str (child_sock, "type", "socket");
-    Jadd_int (child_sock, "req_qty", 4);
-    json_object_object_add (child_sock, "req_child", child_core);
+    Jadd_int (child_sock, "req_qty", 2);
+    Jadd_obj (child_sock, "req_children", ja);
+
+    /* jtago = Jnew (); */
+    /* Jadd_bool (jtago, "maytag", true); */
+    /* Jadd_bool (jtago, "yourtag", true); */
 
     req_res = Jnew ();
     Jadd_str (req_res, "type", "node");
+    /* Jadd_obj (req_res, "tags", jtago); */
     Jadd_int (req_res, "req_qty", 2);
-    json_object_object_add (req_res, "req_child", child_sock);
+    Jadd_obj (req_res, "req_child", child_sock);
 
-    resrc_tree = resrc_phys_tree (resrc);
+    sample_resrc = resrc_new_from_json (req_res, NULL);
+    Jput (req_res);
+    ok ((sample_resrc != NULL), "sample resource composite valid");
+    if (!sample_resrc)
+        goto ret;
+
+    if (verbose) {
+        printf ("Listing sample tree\n");
+        resrc_tree_print (resrc_phys_tree (sample_resrc));
+        printf ("End of sample tree\n");
+    }
+
     init_time();
-    found = resrc_tree_search (resrc_tree_children (resrc_tree), found_res,
-                               req_res, false);
-    if (found) {
-        printf ("Found %d composite resources in %lf\n", found,
-                ((double)get_time())/1000000);
+    found = resrc_tree_search ((resrc_tree_list_t *)resrc_tree_list, found_res,
+                               resrc_phys_tree (sample_resrc), false);
+
+    ok (found, "found %d composite resources in %lf", found,
+        ((double)get_time())/1000000);
+    if (!found)
+        goto ret;
+
+    if (verbose) {
         resrc_id = resrc_list_first (found_res);
         while (resrc_id) {
             printf ("resrc_id %s\n", resrc_id);
             resrc_id = resrc_list_next (found_res);
         }
     }
-    Jput (req_res);
 
     init_time();
     o = resrc_serialize (resrcs, found_res);
-    printf ("Found resource serialization took: %lf\n",
-            ((double)get_time())/1000000);
-    printf ("The found resources serialized: %s\n", Jtostr (o));
+    ok ((o != NULL), "found resource serialization took: %lf",
+        ((double)get_time())/1000000);
+
+    if (verbose) {
+        printf ("The found resources serialized: %s\n", Jtostr (o));
+    }
     Jput (o);
 
     init_time();
-    resrc_allocate_resources (resrcs, found_res, 1);
-    resrc_allocate_resources (resrcs, found_res, 2);
-    resrc_allocate_resources (resrcs, found_res, 3);
-    resrc_reserve_resources (resrcs, found_res, 4);
-    printf ("allocated\n");
-    printf("allocate and reserve took: %lf\n", ((double)get_time())/1000000);
-    /* resrc_print_resources (resrcs); */
+    rc = resrc_allocate_resources (resrcs, found_res, 1);
+    ok (!rc, "successfully allocated resources for job 1");
+    rc = resrc_allocate_resources (resrcs, found_res, 2);
+    ok (!rc, "successfully allocated resources for job 2");
+    rc = resrc_allocate_resources (resrcs, found_res, 3);
+    ok (!rc, "successfully allocated resources for job 3");
+    rc = resrc_reserve_resources (resrcs, found_res, 4);
+    ok (!rc, "successfully reserved resources for job 4");
+
+    printf ("allocate and reserve took: %lf\n", ((double)get_time())/1000000);
+
+    if (verbose) {
+        printf ("Allocated and reserved resources\n");
+        resrc_print_resources (resrcs);
+    }
+
     init_time();
-    resrc_release_resources (resrcs, found_res, 1);
-    printf ("released\n");
-    printf("release took: %lf\n", ((double)get_time())/1000000);
+    rc = resrc_release_resources (resrcs, found_res, 1);
+    ok (!rc, "resource release of job 1 took: %lf",
+        ((double)get_time())/1000000);
+
+    if (verbose) {
+        printf ("Same resources without job 1\n");
+        resrc_print_resources (resrcs);
+    }
+
     init_time();
     resrc_id_list_destroy (found_res);
-    /* resrc_print_resources (resrcs); */
-
     resrc_destroy_resources (&resrcs);
     printf("destroy took: %lf\n", ((double)get_time())/1000000);
-
-    return 0;
+ret:
+    done_testing ();
 }
 
+
+/*
+ * vi: ts=4 sw=4 expandtab
+ */
